requirements.txt
asgiref==3.3.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-appconf==1.0.4
django-googledrive-storage==1.5.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
packaging==20.9
Pillow==8.1.2
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
